back/API/requirements.txt
Flask==3.0.2
flask-cors
flask-restful
mock
firebase_admin
mock-firestore
pycryptodome
PyJWT